+++ src/PSRule.Rules.Azure/rules/Azure.PostgreSQL.Rule.ps1
@@ -28,3 +28,9 @@ Rule 'Azure.PostgreSQL.AllowAzureAccess' -If { ResourceType 'Microsoft.DBforPost
     })
     $firewallRules.Length -eq 0;
 }
+
+# Synopsis: Determine if there is an excessive number of permitted IP addresses
+Rule 'Azure.PostgreSQL.FirewallIPRange' -If { ResourceType 'Microsoft.DBforPostgreSQL/servers' } -Tag @{ severity = 'Important'; category = 'Security configuration' } {
+    $summary = GetIPAddressSummary
+    $summary.Public -le 10;
+}